refactor: add new Subscriber interface file
