+++ types/subscriber.ts
@@ -1,0 +1,8 @@
+export interface Subscriber {
+  id?: number;
+  name: string;
+  email: string;
+  phone: string;
+  rfc: string;
+  spouse: string;
+};
